Replace old interface with new interface.
