+++ CLASSIC_Interface.py
@@ -1,11 +1,13 @@
-# CLASSIC GUI WITH PySide6 (NOW WORKS WITH 3.11!)
 import os
 import sys
 import time
+import io
 import platform
 import subprocess
+import traceback
 import multiprocessing
 import asyncio
+
 try:  # soundfile (specically its Numpy dependency) seem to cause virus alerts from some AV programs, including Windows Defender.
     import soundfile as sfile
     import sounddevice as sdev
@@ -13,187 +15,53 @@ try:  # soundfile (specically its Numpy dependency) seem to cause virus alerts f
 except ImportError:
     has_soundfile = False
 # sfile and sdev need Numpy
+from pathlib import Path
+from PySide6.QtWidgets import (QApplication, QMainWindow, QTabWidget, QWidget, QVBoxLayout, QHBoxLayout,
+                               QPushButton, QLineEdit, QLabel, QFileDialog, QSizePolicy, QMessageBox, QFrame,
+                               QCheckBox, QGridLayout, QTextEdit, QPlainTextEdit, QDialog, QButtonGroup)
+from PySide6.QtGui import QIcon, QDesktopServices, QBrush, QPixmap, QPalette, QKeyEvent, QFont, QPainter, QFontMetrics
+from PySide6.QtCore import Qt, QUrl, QSize, QObject, Signal, QThread, Slot, QTimer, QEvent, QRect
+
 import CLASSIC_Main as CMain
 import CLASSIC_ScanGame as CGame
 import CLASSIC_ScanLogs as CLogs
-from pathlib import Path
-from functools import partial
-from PySide6 import QtCore, QtGui, QtWidgets
-from PySide6.QtCore import QUrl, QTimer, Slot
-from PySide6.QtGui import QDesktopServices, QPixmap, QIcon
-from PySide6.QtWidgets import QDialog, QFileDialog, QSizePolicy, QWidget, QLabel, QHBoxLayout, QVBoxLayout, QPushButton
 
-CMain.configure_logging()
+class ErrorDialog(QDialog):
+    def __init__(self, error_text):
+        super().__init__()
+        self.setWindowTitle("Error")
+        self.setMinimumSize(600, 400)
+        layout = QVBoxLayout(self)
 
-# FONTS CONFIG
-bold_11 = QtGui.QFont()
-bold_11.setPointSize(11)
-bold_11.setBold(True)
-normal_11 = QtGui.QFont()
-normal_11.setPointSize(11)
-bold_09 = QtGui.QFont()
-bold_09.setPointSize(9)
-bold_09.setBold(True)
+        self.text_edit = QPlainTextEdit(self)
+        self.text_edit.setReadOnly(True)
+        self.text_edit.setPlainText(error_text)
+        layout.addWidget(self.text_edit)
 
+        copy_button = QPushButton("Copy to Clipboard", self)
+        copy_button.clicked.connect(self.copy_to_clipboard)
+        layout.addWidget(copy_button)
 
-# ================================================
-# DEFINE WINDOW ELEMENT TEMPLATES HERE
-# ================================================
-def custom_line_box(parent, geometry, object_name, text):
-    line_edit = QtWidgets.QLineEdit(parent)
-    line_edit.setGeometry(geometry)
-    line_edit.setObjectName(object_name)
-    line_edit.setText(text)
-    return line_edit
+    def copy_to_clipboard(self):
+        QApplication.clipboard().setText(self.text_edit.toPlainText())
 
+def show_exception_box(error_text):
+    dialog = ErrorDialog(error_text)
+    dialog.show()
+    dialog.exec()
 
-def custom_push_button(parent, geometry, object_name, text, font, tooltip="", callback=None):
-    button = QtWidgets.QPushButton(parent)
-    button.setObjectName(object_name)
-    button.setGeometry(geometry)
-    button.setToolTip(tooltip)
-    button.setText(text)
-    button.setFont(font)
-    button.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-    button.setStyleSheet("color: white; background: rgba(10, 10, 10, 0.75); border-radius: 10px; border : 1px solid white; font-family: Yu Gothic")
-    if callback:
-        button.clicked.connect(callback)
-    return button
+def custom_excepthook(exc_type, exc_value, exc_traceback):
+    error_text = ''.join(traceback.format_exception(exc_type, exc_value, exc_traceback))
+    print(error_text)  # Still print to console
+    show_exception_box(error_text)
 
+sys.excepthook = custom_excepthook
 
-def custom_frame(parent, geometry, frame_shape, frame_shadow, object_name):
-    frame = QtWidgets.QFrame(parent)
-    frame.setGeometry(geometry)
-    frame.setFrameShape(frame_shape)
-    frame.setFrameShadow(frame_shadow)
-    frame.setObjectName(object_name)
-    return frame
-
-
-def custom_label(parent, geometry, text, font, object_name):
-    label = QtWidgets.QLabel(parent)
-    label.setGeometry(geometry)
-    label.setText(text)
-    label.setFont(font)
-    label.setObjectName(object_name)
-    label.setStyleSheet("color: white; font-family: Yu Gothic")
-    return label
-
-
-def custom_popup_window(parent, title, text, height=250, callback=""):
-    popup_window = QDialog(parent)
-    popup_window.setWindowTitle(title)
-    popup_window.setWindowFlags(QtCore.Qt.FramelessWindowHint)
-    popup_window.setStyleSheet("color: white; background: rgba(10, 10, 10, 1); border : 1px solid black; font-size: 15px")
-    popup_window.setGeometry(15, 300, 620, height)
-
-    layout = QVBoxLayout()
-    label = QLabel(text, popup_window)
-    # label.setAlignment(Qt.AlignTop)
-    label.setWordWrap(True)
-
-    # Create a horizontal layout for buttons
-    button_layout = QHBoxLayout()
-    ok_button = QPushButton("OK")
-    ok_button.setMinimumSize(100, 50)
-    ok_button.setStyleSheet("color: black; background: rgb(45, 237, 138); font-size: 20px; font-weight: bold")
-
-    close_button = QPushButton("Close")
-    close_button.setMinimumSize(100, 50)
-    close_button.setStyleSheet("color: black; background: rgb(240, 63, 40); font-size: 20px; font-weight: bold")
-
-    # Connect button signals to actions
-    if callback:
-        ok_button.clicked.connect(lambda: QDesktopServices.openUrl(QUrl(callback)))
-    else:
-        ok_button.clicked.connect(popup_window.accept)
-    close_button.clicked.connect(popup_window.reject)
-
-    # Add buttons to the horizontal layout
-    button_layout.addWidget(ok_button)
-    button_layout.addWidget(close_button)
-
-    # Add widgets to the main layout
-    layout.addWidget(label)
-    layout.addLayout(button_layout)
-    popup_window.setLayout(layout)
-    return popup_window
-
-
-def custom_text_box(parent, geometry, text):
-    text_browser = QtWidgets.QTextBrowser(parent)
-    text_browser.setGeometry(geometry)
-    text_browser.setObjectName("text_browser")
-    text_browser.setText(text)
-    text_browser.setStyleSheet("color: white; background: rgba(10, 10, 10, 0.75); border-radius: 10px; border : 1px solid white; font-size: 15px")
-    return text_browser
-
-
-def custom_checkbox_widget(parent, pos_x=250, pos_y=250, size=25, check="", label_text="TEST LABEL"):
-    checkbox_widget = QWidget(parent)
-    checkbox_widget.setGeometry(pos_x, pos_y, 200, 50)
-    layout = QHBoxLayout(checkbox_widget)
-
-    # Create QLabel for image & text.
-    image_label = QLabel()
-    pixmap0 = QPixmap("CLASSIC Data/graphics/unchecked.png")
-    pixmap1 = QPixmap("CLASSIC Data/graphics/checked.png")
-
-    image_label.setPixmap(pixmap0)
-    image_label.setFixedSize(size, size)
-    text_label = QLabel(label_text)
-    text_label.setStyleSheet("color: white; font-family: Yu Gothic")
-
-    # Add image & text labels to layout.
-    layout.addWidget(image_label)
-    layout.addWidget(text_label)
-    checkbox_widget.setLayout(layout)
-
-    # Check assigned YAML setting.
-    status = CMain.classic_settings(check)
-    if status:
-        image_label.setPixmap(pixmap1)
-    else:
-        image_label.setPixmap(pixmap0)
-
-    # Toggle assigned YAML setting.
-    def toggle_setting(_):
-        nonlocal check
-        # Toggle between images when label is clicked.
-        if CMain.classic_settings(check):
-            CMain.yaml_settings("CLASSIC Settings.yaml", f"CLASSIC_Settings.{check}", False)
-            image_label.setPixmap(pixmap0)
-        else:
-            CMain.yaml_settings("CLASSIC Settings.yaml", f"CLASSIC_Settings.{check}", True)
-            image_label.setPixmap(pixmap1)
-
-    image_label.mousePressEvent = toggle_setting
-    return checkbox_widget
-
-
-def custom_hover_button(parent, geometry, object_name, text, font, tooltip="", callback=None):
-    hover_button = QtWidgets.QPushButton(parent)
-    hover_button.setObjectName(object_name)
-    hover_button.setGeometry(geometry)
-    hover_button.setToolTip(tooltip)
-    hover_button.setText(text)
-    hover_button.setFont(font)
-    hover_button.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-    hover_button.setStyleSheet("color: white; background: rgba(10, 10, 10, 0.75); border-radius: 10px; border : 1px solid white; font-family: Yu Gothic")
-    if callback:
-        hover_button.clicked.connect(callback)
-
-    def enter_event(_):
-        hover_button.setText("CHANGE GAME")
-
-    def leave_event(_):
-        hover_button.setText(text)
-
-    hover_button.enterEvent = enter_event
-    hover_button.leaveEvent = leave_event
-    hover_button.setMouseTracking(True)
-    return hover_button
-
+def play_sound(sound_file):
+    if has_soundfile:
+        sound, samplerate = sfile.read(f"CLASSIC Data/sounds/{sound_file}") # type: ignore
+        sdev.play(sound, samplerate) # type: ignore
+        sdev.wait() # type: ignore
 
 def papyrus_worker(q, stop_event):
     while not stop_event.is_set():
@@ -201,110 +69,472 @@ def papyrus_worker(q, stop_event):
         q.put(papyrus_result)
         time.sleep(3)
 
+class OutputRedirector(QObject):
+    outputWritten = Signal(str)
 
-def play_sound(sound_file):
-    if not has_soundfile:
-        return
-    sound, samplerate = sfile.read(f"CLASSIC Data/sounds/{sound_file}") # type: ignore
-    sdev.play(sound, samplerate) # type: ignore
-    sdev.wait() # type: ignore
+    def write(self, text):
+        self.outputWritten.emit(str(text))
 
+    def flush(self):
+        pass
 
-# ================================================
-# CLASSIC MAIN WINDOW
-# ================================================
-class UiCLASSICMainWin(QtWidgets.QMainWindow):
+class CrashLogsScanWorker(QObject):
+    finished = Signal()
+
+    @Slot()
+    def run(self):
+        CLogs.crashlogs_scan()
+        play_sound("classic_notify.wav")
+        self.finished.emit()
+
+class GameFilesScanWorker(QObject):
+    finished = Signal()
+
+    @Slot()
+    def run(self):
+        print(CGame.game_combined_result())
+        print(CGame.mods_combined_result())
+        CGame.write_combined_results()
+        play_sound("classic_notify.wav")
+        self.finished.emit()
+
+class MainWindow(QMainWindow):
     def __init__(self):
         super().__init__()
+        self.setWindowTitle(f"Crash Log Auto Scanner & Setup Integrity Checker | {CMain.yaml_settings('CLASSIC Data/databases/CLASSIC Main.yaml', 'CLASSIC_Info.version')}")
+        self.setWindowIcon(QIcon("CLASSIC Data/graphics/CLASSIC.ico"))
+        self.setStyleSheet("font-family: Yu Gothic; font-size: 13px")
+        self.setMinimumSize(700, 950)  # Increase minimum width from 650 to 700
 
-        async def update_check():
-            return await CMain.classic_update_check(quiet=True)
+        # Set up the custom exception handler for the main window
+        self.installEventFilter(self)
 
+        self.central_widget = QWidget()
+        self.setCentralWidget(self.central_widget)
+        self.main_layout = QVBoxLayout(self.central_widget)
+        self.main_layout.setContentsMargins(10, 10, 10, 10)
+        self.main_layout.setSpacing(10)
+
+        self.tab_widget = QTabWidget()
+        self.main_layout.addWidget(self.tab_widget)
+
+        self.main_tab = QWidget()
+        self.backups_tab = QWidget()
+        self.tab_widget.addTab(self.main_tab, "MAIN OPTIONS")
+        self.tab_widget.addTab(self.backups_tab, "FILE BACKUP")
+        self.scan_button_group = QButtonGroup()
+        self.setup_main_tab()
+        self.setup_backups_tab()
+        # In __init__ method, after setting up the main tab:
+        self.initialize_folder_paths()
+        self.setup_output_redirection()
+        self.output_buffer = ""
+        CMain.main_generate_required()
+        # Perform initial update check
         if CMain.classic_settings("Update Check"):
-            self.is_uptodate = asyncio.run(update_check())
-            if not self.is_uptodate:
-                QTimer.singleShot(0, self.update_popup)
+            QTimer.singleShot(0, self.update_popup)
 
-        # MULTIPROCESSING
-        self.papyrus_process = None
-        self.monitoring_flag = multiprocessing.Value('i', 1)  # Shared Flag Variable
+        self.update_check_timer = QTimer()
+        self.update_check_timer.timeout.connect(self.perform_update_check)
+        self.is_update_check_running = False
 
-        # BACKGROUND CONFIG
-        image_path = "CLASSIC Data/graphics/background.png"
-        self.background_label = QtWidgets.QLabel(self)
-        self.setCentralWidget(self.background_label)
-        pixmap = QtGui.QPixmap(image_path)
-        self.background_label.setPixmap(pixmap)
-        self.background_label.setScaledContents(True)
+        # Set up Papyrus monitoring
+        self.result_queue = multiprocessing.Queue()
+        self.worker_stop_event = multiprocessing.Event()
+        self.worker_process = None
+        self.is_worker_running = False
 
-        # ==================== MAIN WINDOW ITEMS =====================
-        # TABS / SCREENS
-        self.RegButton_TabMain = custom_push_button(self, QtCore.QRect(0, 0, 215, 48), "RegButton_TabMain", "MAIN OPTIONS", bold_11, "")
-        self.RegButton_TabMain.setStyleSheet("color: white; background: rgba(25, 25, 25, 0.90); border-radius: 0px; border : 2px solid white; font-size: 15px")
+        # Initialize thread attributes
+        self.crash_logs_thread = None
+        self.game_files_thread = None
 
-        self.RegButton_TabBackups = custom_push_button(self, QtCore.QRect(215, 0, 215, 48), "RegButton_TabBackups", "FILE BACKUP", bold_11, "", self.open_tab_backups)
-        self.RegButton_TabBackups.setStyleSheet("color: white; background: rgba(10, 10, 10, 0.90); border-radius: 0px; border : 2px dashed white; font-size: 15px")
+        self.timer = QTimer()
+        self.timer.timeout.connect(self.update_output_text_box_papyrus_watcher)
+        self.timer.start(5000)  # Update every 5 seconds
 
-        self.DropDown_GameSelect = custom_hover_button(self, QtCore.QRect(430, 0, 220, 48), "DropDown_GameSelect", "FALLOUT 4", bold_11, "")
-        self.DropDown_GameSelect.setStyleSheet("color: white; background: rgba(240, 63, 40, 0.75); border-radius: 0px; border : 2px solid white; font-size: 15px")
+    def eventFilter(self, watched: QObject, event: QEvent) -> bool:
+        if event.type() == QEvent.KeyPress:
+            key_event = QKeyEvent(event)
+            if key_event.key() == Qt.Key_F12:
+                # Simulate an exception when F12 is pressed (for testing)
+                raise Exception("This is a test exception")
+        return super().eventFilter(watched, event)
 
-        # TOP
+    def update_popup(self):
+        if not self.is_update_check_running:
+            self.is_update_check_running = True
+            self.update_check_timer.start(0)  # Start immediately
 
-        # SEPARATOR STAGING MODS FOLDER
-        self.LBL_ModsFolder = custom_label(self, QtCore.QRect(20, 80, 260, 20), "STAGING MODS FOLDER", bold_11, "LBL_ModsFolder")
-        self.Line_Sep_Mods = custom_frame(self, QtCore.QRect(30, 130, 590, 20), QtWidgets.QFrame.Shape.HLine, QtWidgets.QFrame.Shadow.Sunken, "Line_Sep_Mods")
-        # BROWSE STAGING MODS FOLDER
-        self.Box_SelectedMods = custom_line_box(self, QtCore.QRect(20, 100, 450, 24), "Box_SelectedMods", "(Optional) Press *Browse Folder* to set your staging mods folder location.")
-        self.Box_SelectedMods.setStyleSheet("color: rgb(100, 100, 100)")
-        self.RegButton_BrowseMods = custom_push_button(self, QtCore.QRect(490, 100, 130, 24), "RegButton_BrowseMods", "Browse Folder", normal_11, "", self.select_folder_mods)
+    def perform_update_check(self):
+        self.update_check_timer.stop()
+        asyncio.ensure_future(self.async_update_check())
 
-        # SEPARATOR CUSTOM SCAN FOLDER
-        self.LBL_ScanFolder = custom_label(self, QtCore.QRect(20, 155, 260, 20), "CUSTOM SCAN FOLDER", bold_11, "LBL_ScanFolder")
-        self.Line_Sep_Scan = custom_frame(self, QtCore.QRect(30, 205, 590, 20), QtWidgets.QFrame.Shape.HLine, QtWidgets.QFrame.Shadow.Sunken, "Line_Sep_Scan")
-        # BROWSE CUSTOM SCAN FOLDER
-        self.Box_SelectedScan = custom_line_box(self, QtCore.QRect(20, 175, 450, 24), "Box_SelectedScan", "(Optional) Press *Browse Folder* to set a different scan folder location.")
-        self.Box_SelectedScan.setStyleSheet("color: rgb(100, 100, 100)")
-        self.RegButton_BrowseScan = custom_push_button(self, QtCore.QRect(490, 175, 130, 24), "RegButton_BrowseScan", "Browse Folder", normal_11, "", self.select_folder_scan)
+    async def async_update_check(self):
+        try:
+            is_up_to_date = await CMain.classic_update_check(quiet=True)
+            self.show_update_result(is_up_to_date)
+        except Exception as e:
+            self.show_update_error(str(e))
+        finally:
+            self.is_update_check_running = False
 
-        # TOP MAIN ROW
-        self.RegButton_SCAN_LOGS = custom_push_button(self, QtCore.QRect(35, 245, 270, 48), "RegButton_SCAN_LOGS", "SCAN CRASH LOGS", bold_11, "", self.crash_logs_scan)
-        self.RegButton_SCAN_LOGS.setStyleSheet("color: black; background: rgba(250, 250, 250, 0.90); border-radius: 10px; border : 1px solid white; font-size: 17px")
-        self.RegButton_SCAN_FILES = custom_push_button(self, QtCore.QRect(345, 245, 270, 48), "RegButton_SCAN_FILES", "SCAN GAME FILES", bold_11, "", self.game_files_scan)
-        self.RegButton_SCAN_FILES.setStyleSheet("color: black; background: rgba(250, 250, 250, 0.90); border-radius: 10px; border : 1px solid white; font-size: 17px")
+    def show_update_result(self, is_up_to_date):
+        if is_up_to_date:
+            QMessageBox.information(self, "CLASSIC UPDATE", "You have the latest version of CLASSIC!")
+        else:
+            update_popup_text = CMain.yaml_settings("CLASSIC Data/databases/CLASSIC Main.yaml", "CLASSIC_Interface.update_popup_text")
+            result = QMessageBox.question(self, "CLASSIC UPDATE", update_popup_text,
+                                          QMessageBox.Yes | QMessageBox.No)
+            if result == QMessageBox.Yes:
+                QDesktopServices.openUrl(QUrl("https://github.com/evildarkarchon/CLASSIC-Fallout4/releases/latest"))
 
-        # BOTTOM MAIN ROW
-        self.RegButton_ChangeINI = custom_push_button(self, QtCore.QRect(35, 310, 150, 32), "RegButton_ChangeINI", "CHANGE INI PATH", bold_09, "Select folder of the game's main INI files so CLASSIC can use that location.", self.select_folder_ini)
-        self.RegButton_OpenSettings = custom_push_button(self, QtCore.QRect(220, 310, 210, 32), "RegButton_OpenSettings", "OPEN CLASSIC SETTINGS", bold_09, "Open the CLASSIC Settings.yaml file in your default text editor.", self.open_settings)
-        self.RegButton_CheckUpdates = custom_push_button(self, QtCore.QRect(465, 310, 150, 32), "RegButton_CheckUpdates", "CHECK UPDATES", bold_09, "Check for new CLASSIC versions (CLASSIC does this automatically every 7 days).", self.update_popup)
+    def show_update_error(self, error_message):
+        QMessageBox.warning(self, "Update Check Failed", f"Failed to check for updates: {error_message}")
 
-        # CHECK EXISTING BROWSE PATHS
-        SCAN_folder = CMain.classic_settings("SCAN Custom Path")
-        if SCAN_folder:
-            self.Box_SelectedScan.setText(SCAN_folder.strip())
-            self.Box_SelectedScan.setStyleSheet("color: black")
-        MODS_folder = CMain.classic_settings("MODS Folder Path")
-        if MODS_folder:
-            self.Box_SelectedMods.setText(MODS_folder.strip())
-            self.Box_SelectedMods.setStyleSheet("color: black")
+    def setup_main_tab(self):
+        layout = QVBoxLayout(self.main_tab)
+        layout.setContentsMargins(20, 10, 20, 10)
+        layout.setSpacing(10)
 
-        # SEGMENT - SETTINGS
-        self.Line_Sep_Settings = custom_frame(self, QtCore.QRect(30, 360, 590, 20), QtWidgets.QFrame.Shape.HLine, QtWidgets.QFrame.Shadow.Sunken, "Line_Sep_Settings")
-        self.LBL_Settings = custom_label(self, QtCore.QRect(260, 380, 180, 16), "CLASSIC SETTINGS", bold_11, "LBL_Settings")
-        # CHECKBOXES - Column 1
-        self.ChkBT_FCXMode = custom_checkbox_widget(self, pos_x=70, pos_y=400, label_text="FCX MODE", check="FCX Mode")
-        self.ChkBT_VRMode = custom_checkbox_widget(self, pos_x=70, pos_y=450, label_text="VR MODE", check="VR Mode")
-        # CHECKBOXES - Column 2
-        self.ChkBT_SimpleLogs = custom_checkbox_widget(self, pos_x=250, pos_y=400, label_text="SIMPLIFY LOGS", check="Simplify Logs")
-        self.ChkBT_ShowFormID = custom_checkbox_widget(self, pos_x=250, pos_y=450, label_text="SHOW FID VALUES", check="Show FormID Values")
-        # CHECKBOXES - Column 3
-        self.ChkBT_Update = custom_checkbox_widget(self, pos_x=430, pos_y=400, label_text="UPDATE CHECK", check="Update Check")
-        self.ChkBT_Unsolved = custom_checkbox_widget(self, pos_x=430, pos_y=450, label_text="MOVE INVALID LOGS", check="Move Unsolved Logs")
+        # Top section
+        self.mods_folder_edit = self.setup_folder_section(layout, "STAGING MODS FOLDER", "Box_SelectedMods", self.select_folder_mods)
+        self.scan_folder_edit = self.setup_folder_section(layout, "CUSTOM SCAN FOLDER", "Box_SelectedScan", self.select_folder_scan)
 
-        # SEPARATOR WEBSITE LINKS
-        self.Line_Sep_Links = custom_frame(self, QtCore.QRect(30, 510, 590, 20), QtWidgets.QFrame.Shape.HLine, QtWidgets.QFrame.Shadow.Sunken, "Line_Sep_Links")
-        self.LBL_ArtWeb = custom_label(self, QtCore.QRect(200, 530, 590, 20), "ARTICLES / WEBSITES / NEXUS LINKS", bold_11, "LBL_ArtWeb")
+        # Add first separator
+        layout.addWidget(self.create_separator())
 
-        # Articles & Websites - ADD LINK BUTTONS FOR BETHINI, DDS SCANNER & WRYE BASH
+        # Main buttons section
+        self.setup_main_buttons(layout)
+
+        # Add second separator
+        layout.addWidget(self.create_separator())
+
+        # Checkbox section
+        self.setup_checkboxes(layout)
+
+        # Articles section
+        self.setup_articles_section(layout)
+
+        # Add a separator before bottom buttons
+        layout.addWidget(self.create_separator())
+
+        # Bottom buttons
+        self.setup_bottom_buttons(layout)
+
+        # Add output text box
+        self.setup_output_text_box(layout)
+
+        # Add some spacing
+        layout.addSpacing(10)
+
+        # Set the layout to be stretchable
+        layout.setStretchFactor(self.output_text_box, 1)
+
+    def setup_backups_tab(self):
+        layout = QVBoxLayout(self.backups_tab)
+        layout.setContentsMargins(20, 10, 20, 10)
+        layout.setSpacing(10)
+
+        # Add explanation labels
+        layout.addWidget(QLabel("BACKUP > Backup files from the game folder into the CLASSIC Backup folder."))
+        layout.addWidget(QLabel("RESTORE > Restore file backup from the CLASSIC Backup folder into the game folder."))
+        layout.addWidget(QLabel("REMOVE > Remove files only from the game folder without removing existing backups."))
+
+        # Add separators and category buttons
+        categories = ["XSE", "RESHADE", "VULKAN", "ENB"]
+        for category in categories:
+            layout.addWidget(self.create_separator())
+            layout.addWidget(QLabel(category, alignment=Qt.AlignCenter))
+
+            button_layout = QHBoxLayout()
+
+            backup_button = QPushButton(f"BACKUP {category}")
+            backup_button.clicked.connect(lambda _, c=category: self.classic_files_manage(f"Backup {c}", "BACKUP"))
+            button_layout.addWidget(backup_button)
+
+            restore_button = QPushButton(f"RESTORE {category}")
+            restore_button.clicked.connect(lambda _, c=category: self.classic_files_manage(f"Backup {c}", "RESTORE"))
+            restore_button.setEnabled(False)  # Initially disabled
+            setattr(self, f"RestoreButton_{category}", restore_button)  # Store reference to the button
+            button_layout.addWidget(restore_button)
+
+            remove_button = QPushButton(f"REMOVE {category}")
+            remove_button.clicked.connect(lambda _, c=category: self.classic_files_manage(f"Backup {c}", "REMOVE"))
+            button_layout.addWidget(remove_button)
+
+            layout.addLayout(button_layout)
+
+        # Check if backups exist and enable restore buttons accordingly
+        self.check_existing_backups()
+
+        # Add a button to open the backups folder
+        open_backups_button = QPushButton("OPEN CLASSIC BACKUPS")
+        open_backups_button.clicked.connect(self.open_backup_folder)
+        layout.addWidget(open_backups_button)
+
+    def check_existing_backups(self):
+        for category in ["XSE", "RESHADE", "VULKAN", "ENB"]:
+            backup_path = f"CLASSIC Backup/Game Files/Backup {category}"
+            if os.path.isdir(backup_path) and any(Path(backup_path).iterdir()):
+                restore_button = getattr(self, f"RestoreButton_{category}", None)
+                if restore_button:
+                    restore_button.setEnabled(True)
+                    restore_button.setStyleSheet("""
+                        QPushButton {
+                            color: black;
+                            background: rgb(250, 250, 250);
+                            border-radius: 10px;
+                            border: 2px solid black;
+                        }
+                    """)
+
+    def add_backup_section(self, layout, title, backup_type):
+        layout.addWidget(self.create_separator())
+
+        title_label = QLabel(title)
+        title_label.setAlignment(Qt.AlignCenter)
+        title_label.setStyleSheet("color: white; font-weight: bold; font-size: 14px;")
+        layout.addWidget(title_label)
+
+        buttons_layout = QHBoxLayout()
+        backup_button = QPushButton(f"BACKUP {backup_type}")
+        restore_button = QPushButton(f"RESTORE {backup_type}")
+        remove_button = QPushButton(f"REMOVE {backup_type}")
+
+        for button, action in [(backup_button, "BACKUP"), (restore_button, "RESTORE"), (remove_button, "REMOVE")]:
+            button.clicked.connect(lambda _, b=backup_type, a=action: self.classic_files_manage(f"Backup {b}", a))
+            button.setStyleSheet("""
+                QPushButton {
+                    color: white;
+                    background: rgba(10, 10, 10, 0.75);
+                    border-radius: 10px;
+                    border: 1px solid white;
+                    font-size: 11px;
+                    min-height: 48px;
+                    max-height: 48px;
+                    min-width: 180px;
+                    max-width: 180px;
+                }
+            """)
+            buttons_layout.addWidget(button)
+
+        layout.addLayout(buttons_layout)
+
+    def classic_files_manage(self, selected_list, selected_mode="BACKUP"):
+        list_name = selected_list.split(" ", 1)
+        try:
+            CGame.game_files_manage(selected_list, selected_mode)
+            if selected_mode == "BACKUP":
+                # Enable the corresponding restore button
+                restore_button = getattr(self, f"RestoreButton_{list_name[1]}", None)
+                if restore_button:
+                    restore_button.setEnabled(True)
+                    restore_button.setStyleSheet("""
+                        QPushButton {
+                            color: black;
+                            background: rgb(250, 250, 250);
+                            border-radius: 10px;
+                            border: 2px solid black;
+                        }
+                    """)
+        except PermissionError:
+            QMessageBox.critical(self, "Error", "Unable to access files from your game folder. Please run CLASSIC in admin mode to resolve this problem.")
+
+    def help_popup_backup(self):
+        help_popup_text = CMain.yaml_settings("CLASSIC Data/databases/CLASSIC Main.yaml", "CLASSIC_Interface.help_popup_backup")
+        QMessageBox.information(self, "NEED HELP?", help_popup_text)
+
+    @staticmethod
+    def open_backup_folder():
+        backup_path = os.path.join(os.getcwd(), "CLASSIC Backup", "Game Files")
+        QDesktopServices.openUrl(QUrl.fromLocalFile(backup_path))
+
+    def setup_output_text_box(self, layout):
+        self.output_text_box = QTextEdit(self)
+        self.output_text_box.setReadOnly(True)
+        self.output_text_box.setStyleSheet("""
+            QTextEdit {
+                color: white;
+                font-family: Consolas, monospace;
+                background: rgba(10, 10, 10, 0.75);
+                border-radius: 10px;
+                border: 1px solid white;
+                font-size: 13px;
+            }
+        """) # Have to use alternate font here because the default font doesn't support some characters.
+
+        self.output_text_box.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
+        self.output_text_box.setMinimumHeight(150)
+        layout.addWidget(self.output_text_box)
+
+        self.output_buffer = ""
+
+    def update_output_text_box(self, text):
+        try:
+            if isinstance(text, bytes):
+                text = text.decode('utf-8', errors='replace')
+            else:
+                text = str(text)
+
+            self.output_buffer += text
+            lines = self.output_buffer.splitlines(True)
+
+            complete_lines = lines[:-1] if not self.output_buffer.endswith('\n') else lines
+            if complete_lines:
+                current_text = self.output_text_box.toPlainText()
+                new_text = current_text + ''.join(complete_lines)
+                self.output_text_box.setPlainText(new_text)
+
+                # Scroll to the bottom
+                scrollbar = self.output_text_box.verticalScrollBar()
+                scrollbar.setValue(scrollbar.maximum())
+
+            self.output_buffer = lines[-1] if not self.output_buffer.endswith('\n') else ""
+        except Exception as e:
+            print(f"Error in update_output_text_box: {e}")
+
+    def process_lines(self, lines):
+        for line in lines:
+            stripped_line = line.rstrip()
+            if stripped_line or line.endswith('\n'):
+                self.output_text_box.append(stripped_line)
+
+        # Scroll to the bottom of the text box
+        scrollbar = self.output_text_box.verticalScrollBar()
+        scrollbar.setValue(scrollbar.maximum())
+
+    def setup_output_redirection(self):
+        self.output_redirector = OutputRedirector()
+        self.output_redirector.outputWritten.connect(self.update_output_text_box)
+        sys.stdout = self.output_redirector
+        sys.stderr = self.output_redirector  # Also redirect stderr
+
+    @staticmethod
+    def create_separator():
+        separator = QFrame()
+        separator.setFrameShape(QFrame.HLine)
+        separator.setFrameShadow(QFrame.Sunken)
+        return separator
+
+    def setup_checkboxes(self, layout):
+        checkbox_layout = QVBoxLayout()
+
+        # Title for the checkbox section
+        title_label = QLabel("CLASSIC SETTINGS")
+        title_label.setAlignment(Qt.AlignCenter)
+        title_label.setStyleSheet("font-weight: bold; font-size: 14px;")
+        checkbox_layout.addWidget(title_label)
+
+        # Grid for checkboxes
+        grid_layout = QGridLayout()
+        grid_layout.setHorizontalSpacing(40)  # Increased spacing
+        grid_layout.setVerticalSpacing(20)  # Increased spacing
+
+        checkboxes = [
+            ("FCX MODE", "FCX Mode"),
+            ("SIMPLIFY LOGS", "Simplify Logs"),
+            ("UPDATE CHECK", "Update Check"),
+            ("VR MODE", "VR Mode"),
+            ("SHOW FID VALUES", "Show FormID Values"),
+            ("MOVE INVALID LOGS", "Move Unsolved Logs")
+        ]
+
+        for index, (label, setting) in enumerate(checkboxes):
+            checkbox = self.create_checkbox(label, setting)
+            row = index // 3
+            col = index % 3
+            grid_layout.addWidget(checkbox, row, col, Qt.AlignLeft)
+
+        checkbox_layout.addLayout(grid_layout)
+
+        # Add some vertical spacing
+        checkbox_layout.addSpacing(20)
+
+        layout.addLayout(checkbox_layout)
+
+        # Add a separator after the checkboxes
+        layout.addWidget(self.create_separator())
+
+    @staticmethod
+    def create_checkbox(label_text, setting):
+        checkbox = QCheckBox(label_text)
+        checkbox.setChecked(CMain.classic_settings(setting))
+        checkbox.stateChanged.connect(lambda state: CMain.yaml_settings("CLASSIC Settings.yaml", f"CLASSIC_Settings.{setting}", bool(state)))
+
+        # Apply custom style sheet
+        checkbox.setStyleSheet("""
+            QCheckBox {
+                spacing: 10px;
+            }
+            QCheckBox::indicator {
+                width: 25px;
+                height: 25px;
+            }
+            QCheckBox::indicator:unchecked {
+                image: url(CLASSIC Data/graphics/unchecked.png);
+            }
+            QCheckBox::indicator:checked {
+                image: url(CLASSIC Data/graphics/checked.png);
+            }
+        """)
+
+        return checkbox
+
+    @staticmethod
+    def setup_folder_section(layout, title, box_name, browse_callback, tooltip=""):
+        section_layout = QHBoxLayout()
+        section_layout.setContentsMargins(0, 0, 0, 0)
+        section_layout.setSpacing(5)
+
+        label = QLabel(title)
+        label.setAlignment(Qt.AlignLeft | Qt.AlignVCenter)
+        label.setFixedWidth(180)
+        section_layout.addWidget(label)
+
+        line_edit = QLineEdit()
+        line_edit.setObjectName(box_name)
+        section_layout.addWidget(line_edit, 1)
+
+        browse_button = QPushButton("Browse Folder")
+        if tooltip:
+            browse_button.setToolTip(tooltip)
+        browse_button.clicked.connect(browse_callback)
+        section_layout.addWidget(browse_button)
+
+        layout.addLayout(section_layout)
+        return line_edit  # Return the created QLineEdit
+
+    def setup_main_buttons(self, layout):
+        # Main action buttons
+        main_buttons_layout = QHBoxLayout()
+        main_buttons_layout.setSpacing(10)
+        self.crash_logs_button = self.add_main_button(main_buttons_layout, "SCAN CRASH LOGS", self.crash_logs_scan)
+        self.scan_button_group.addButton(self.crash_logs_button)
+        self.game_files_button = self.add_main_button(main_buttons_layout, "SCAN GAME FILES", self.game_files_scan)
+        self.scan_button_group.addButton(self.game_files_button)
+        layout.addLayout(main_buttons_layout)
+
+        # Bottom row buttons
+        bottom_buttons_layout = QHBoxLayout()
+        bottom_buttons_layout.setSpacing(5)
+        self.add_bottom_button(bottom_buttons_layout, "CHANGE INI PATH", self.select_folder_ini)
+        self.add_bottom_button(bottom_buttons_layout, "OPEN CLASSIC SETTINGS", self.open_settings)
+        self.add_bottom_button(bottom_buttons_layout, "CHECK UPDATES", self.update_popup)
+        layout.addLayout(bottom_buttons_layout)
+
+    @staticmethod
+    def setup_articles_section(layout):
+        # Title for the articles section
+        title_label = QLabel("ARTICLES / WEBSITES / NEXUS LINKS")
+        title_label.setAlignment(Qt.AlignCenter)
+        title_label.setStyleSheet("font-weight: bold; font-size: 14px;")
+        layout.addWidget(title_label)
+
+        # Grid layout for article buttons
+        grid_layout = QGridLayout()
+        grid_layout.setHorizontalSpacing(10)
+        grid_layout.setVerticalSpacing(10)
+
         button_data = [
             {"text": "BUFFOUT 4 INSTALLATION", "url": "https://www.nexusmods.com/fallout4/articles/3115"},
             {"text": "FALLOUT 4 SETUP TIPS", "url": "https://www.nexusmods.com/fallout4/articles/4141"},
@@ -317,302 +547,291 @@ class UiCLASSICMainWin(QtWidgets.QMainWindow):
             {"text": "WRYE BASH TOOL", "url": "https://www.nexusmods.com/fallout4/mods/20032"}
         ]
 
-        # ARRANGE BUTTONS IN GRID
         for i, data in enumerate(button_data):
-            button = QtWidgets.QPushButton(self)
-            button.setGeometry(QtCore.QRect(45 + i % 3 * 190, 570 + i // 3 * 60, 180, 50))
-            button.setObjectName("ArtBT_" + data["text"].replace(" ", ""))
-            button.setText(data["text"])
-            button.setStyleSheet("color: white; border-radius: 5px; border : 1px solid white; font-size: 11px; font-weight: bold")
-            open_url = partial(QDesktopServices.openUrl, QUrl(data["url"]))
-            button.clicked.connect(open_url)
+            button = QPushButton(data["text"])
+            button.setFixedSize(180, 50)  # Set fixed size for buttons
+            button.setStyleSheet("""
+                QPushButton {
+                    color: white;
+                    background-color: rgba(10, 10, 10, 0.75);
+                    border: 1px solid white;
+                    border-radius: 5px;
+                    padding: 5px;
+                    font-size: 11px;
+                    font-weight: bold;
+                }
+                QPushButton:hover {
+                    background-color: rgba(50, 50, 50, 0.75);
+                }
+                QPushButton:disabled {
+                    color: gray;
+                    background-color: rgba(10, 10, 10, 0.75);
+                }
+            """)
+            button.clicked.connect(lambda _, url=data["url"]: QDesktopServices.openUrl(QUrl(url)))
+            row = i // 3
+            col = i % 3
+            grid_layout.addWidget(button, row, col, Qt.AlignCenter)
 
-        # BOTTOM
+        layout.addLayout(grid_layout)
 
-        # Button - HELP
-        self.RegButton_Help = custom_push_button(self, QtCore.QRect(20, 770, 110, 30), "RegButton_Help", "HELP", normal_11, "", self.help_popup_main)
-        # Button - PAPYRUS MONITORING
-        self.RegButton_Papyrus = custom_push_button(self, QtCore.QRect(195, 770, 260, 30), "RegButton_Papyrus", "START PAPYRUS MONITORING", bold_11, "", self.toggle_papyrus_worker)
-        self.RegButton_Papyrus.setStyleSheet("color: black; background: rgb(45, 237, 138); border-radius: 10px; border : 1px solid black")
-        # Button - EXIT
-        self.RegButton_Exit = custom_push_button(self, QtCore.QRect(520, 770, 110, 30), "RegButton_Exit", "EXIT", normal_11, "", lambda: QtWidgets.QApplication.quit())
-        # Text Box - SHARED
-        self.TXT_Window = custom_text_box(self, QtCore.QRect(20, 810, 610, 120), "Crash Log Auto Scanner & Setup Integrity Checker | Made by: Poet \nContributors: evildarkarchon | kittivelae | AtomicFallout757")
+        # Add some vertical spacing after the articles section
+        layout.addSpacing(20)
 
-        QtCore.QMetaObject.connectSlotsByName(self)
+    def setup_bottom_buttons(self, layout):
+        bottom_layout = QHBoxLayout()
+        bottom_layout.setSpacing(5)  # Reduce spacing between buttons
 
-        # ================= PAPYRUS WORKER =================
-        self.result_queue = multiprocessing.Queue()
-        self.worker_stop_event = multiprocessing.Event()
-        self.worker_process = None
+        # ABOUT button
+        about_button = QPushButton("ABOUT")
+        about_button.setFixedSize(80, 30)  # Reduce width
+        about_button.clicked.connect(self.show_about)
+        about_button.setStyleSheet("""
+            QPushButton {
+                color: white;
+                background: rgba(10, 10, 10, 0.75);
+                border-radius: 10px;
+                border: 1px solid white;
+                font-size: 11px;
+            }
+        """)
+        bottom_layout.addWidget(about_button)
 
-        self.timer = QTimer()
-        # noinspection PyUnresolvedReferences
-        self.timer.timeout.connect(self.update_text_window)
-        self.timer.start(5000)  # Update every 5 seconds.
+        # HELP button
+        help_button = QPushButton("HELP")
+        help_button.setFixedSize(80, 30)  # Reduce width
+        help_button.clicked.connect(self.help_popup_main)
+        help_button.setStyleSheet("""
+            QPushButton {
+                color: white;
+                background: rgba(10, 10, 10, 0.75);
+                border-radius: 10px;
+                border: 1px solid white;
+                font-size: 11px;
+            }
+        """)
+        bottom_layout.addWidget(help_button)
 
-        self.is_worker_running = False
+        # PAPYRUS MONITORING button
+        self.papyrus_button = QPushButton("START PAPYRUS MONITORING")
+        self.papyrus_button.setFixedHeight(30)
+        self.papyrus_button.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Fixed)
+        self.papyrus_button.clicked.connect(self.toggle_papyrus_worker)
+        self.papyrus_button.setStyleSheet("""
+            QPushButton {
+                color: black;
+                background: rgb(45, 237, 138);
+                border-radius: 10px;
+                border: 1px solid black;
+                font-weight: bold;
+                font-size: 11px;
+            }
+        """)
+        bottom_layout.addWidget(self.papyrus_button)
 
-    # ================= MAIN BUTTON FUNCTIONS ===================
-    # @staticmethod recommended for func that don't call "self".
+        # EXIT button
+        exit_button = QPushButton("EXIT")
+        exit_button.setFixedSize(80, 30)  # Reduce width
+        exit_button.clicked.connect(QApplication.quit)
+        exit_button.setStyleSheet("""
+            QPushButton {
+                color: white;
+                background: rgba(10, 10, 10, 0.75);
+                border-radius: 10px;
+                border: 1px solid white;
+                font-size: 11px;
+            }
+        """)
+        bottom_layout.addWidget(exit_button)
+
+        layout.addLayout(bottom_layout)
+
+    def show_about(self):
+        about_text = ("Crash Log Auto Scanner & Setup Integrity Checker\n\n"
+                    "Made by: Poet\n"
+                    "Contributors: evildarkarchon | kittivelae | AtomicFallout757")
+        QMessageBox.about(self, "About CLASSIC", about_text)
+
+    def help_popup_main(self):
+        help_popup_text = CMain.yaml_settings("CLASSIC Data/databases/CLASSIC Main.yaml", "CLASSIC_Interface.help_popup_main")
+        QMessageBox.information(self, "NEED HELP?", help_popup_text)
 
     @staticmethod
-    def open_tab_backups():
-        screen_backups = UiCLASSICBackups()
-        screen_switch.addWidget(screen_backups)
-        screen_switch.setCurrentIndex(screen_switch.currentIndex() + 1)
+    def add_main_button(layout, text, callback, tooltip=""):
+        button = QPushButton(text)
+        button.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Fixed)
+        button.setStyleSheet("""
+            QPushButton {
+                color: black;
+                background: rgba(250, 250, 250, 0.90);
+                border-radius: 10px;
+                border: 1px solid white;
+                font-size: 17px;
+                font-weight: bold;  /* Add this line to make the text bold */
+                min-height: 48px;
+                max-height: 48px;
+            }
+            QPushButton:disabled {
+                color: gray;
+                background-color: rgba(10, 10, 10, 0.75);
+            }
+        """)
+        if tooltip:
+            button.setToolTip(tooltip)
+        button.clicked.connect(callback)
+        layout.addWidget(button)
+        return button
 
     @staticmethod
-    def crash_logs_scan():
-        CLogs.crashlogs_scan()
-        play_sound("classic_notify.wav")
+    def add_bottom_button(layout, text, callback, tooltip=""):
+        button = QPushButton(text)
+        button.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Fixed)
+        button.setStyleSheet("""
+            color: white;
+            background: rgba(10, 10, 10, 0.75);
+            border-radius: 10px;
+            border: 1px solid white;
+            font-size: 11px;
+            min-height: 32px;
+            max-height: 32px;
+        """)
+        if tooltip:
+            button.setToolTip(tooltip)
+        button.clicked.connect(callback)
+        layout.addWidget(button)
 
+    def select_folder_scan(self):
+        folder = QFileDialog.getExistingDirectory(self, "Select Custom Scan Folder")
+        if folder:
+            self.scan_folder_edit.setText(folder)
+            CMain.yaml_settings("CLASSIC Settings.yaml", "CLASSIC_Settings.SCAN Custom Path", folder)
+
+    def select_folder_mods(self):
+        folder = QFileDialog.getExistingDirectory(self, "Select Staging Mods Folder")
+        if folder:
+            self.mods_folder_edit.setText(folder)
+            CMain.yaml_settings("CLASSIC Settings.yaml", "CLASSIC_Settings.MODS Folder Path", folder)
+
+    def initialize_folder_paths(self):
+        scan_folder = CMain.classic_settings("SCAN Custom Path")
+        mods_folder = CMain.classic_settings("MODS Folder Path")
+
+        if scan_folder:
+            self.scan_folder_edit.setText(scan_folder)
+        if mods_folder:
+            self.mods_folder_edit.setText(mods_folder)
+
+    def select_folder_ini(self):
+        folder = QFileDialog.getExistingDirectory(self)
+        if folder:
+            CMain.yaml_settings("CLASSIC Settings.yaml", "CLASSIC_Settings.INI Folder Path", folder)
+            QMessageBox.information(self, "New INI Path Set", f"You have set the new path to: \n{folder}")
     @staticmethod
-    def game_files_scan():
-        print(CGame.game_combined_result())
-        print(CGame.mods_combined_result())
-        CGame.write_combined_results()
-        play_sound("classic_notify.wav")
+    def open_settings():
+        settings_file = "CLASSIC Settings.yaml"
+        QDesktopServices.openUrl(QUrl.fromLocalFile(settings_file))
 
-    @Slot()
+    def crash_logs_scan(self):
+        if self.crash_logs_thread is None:
+            self.crash_logs_thread = QThread()
+            self.crash_logs_worker = CrashLogsScanWorker()
+            self.crash_logs_worker.moveToThread(self.crash_logs_thread)
+            self.crash_logs_thread.started.connect(self.crash_logs_worker.run)
+            self.crash_logs_worker.finished.connect(self.crash_logs_thread.quit)
+            self.crash_logs_worker.finished.connect(self.crash_logs_worker.deleteLater)
+            self.crash_logs_thread.finished.connect(self.crash_logs_thread.deleteLater)
+            self.crash_logs_thread.finished.connect(self.crash_logs_scan_finished)
+
+            # Disable buttons and update text
+            self.disable_scan_buttons()
+
+            self.crash_logs_thread.start()
+
+    def game_files_scan(self):
+        if self.game_files_thread is None:
+            self.game_files_thread = QThread()
+            self.game_files_worker = GameFilesScanWorker()
+            self.game_files_worker.moveToThread(self.game_files_thread)
+            self.game_files_thread.started.connect(self.game_files_worker.run)
+            self.game_files_worker.finished.connect(self.game_files_thread.quit)
+            self.game_files_worker.finished.connect(self.game_files_worker.deleteLater)
+            self.game_files_thread.finished.connect(self.game_files_thread.deleteLater)
+            self.game_files_thread.finished.connect(self.game_files_scan_finished)
+
+            # Disable buttons and update text
+            self.disable_scan_buttons()
+
+            self.game_files_thread.start()
+
+    def disable_scan_buttons(self):
+        for button_id in self.scan_button_group.buttons():
+            button_id.setEnabled(False)
+
+    def enable_scan_buttons(self):
+        for button_id in self.scan_button_group.buttons():
+            button_id.setEnabled(True)
+
+    def crash_logs_scan_finished(self):
+        self.crash_logs_thread = None
+        self.enable_scan_buttons()
+        self.update_output_text_box("Crash logs scan completed.")
+
+    def game_files_scan_finished(self):
+        self.game_files_thread = None
+        self.enable_scan_buttons()
+        self.update_output_text_box("Game files scan completed.")
+
     def toggle_papyrus_worker(self):
         if not self.is_worker_running:
             self.worker_stop_event.clear()
             self.worker_process = multiprocessing.Process(target=papyrus_worker, args=(self.result_queue, self.worker_stop_event))
             self.worker_process.daemon = True
             self.worker_process.start()
-            self.RegButton_Papyrus.setText("STOP PAPYRUS MONITORING")
-            self.RegButton_Papyrus.setStyleSheet("background: rgb(240, 63, 40); border-radius: 10px")
+            self.papyrus_button.setText("STOP PAPYRUS MONITORING")
+            self.papyrus_button.setStyleSheet("""
+                QPushButton {
+                    color: black;
+                    background: rgb(240, 63, 40);
+                    border-radius: 10px;
+                    border: 1px solid black;
+                    font-weight: bold;
+                }
+            """)
         else:
             self.worker_stop_event.set()
-            self.worker_process.join()
+            if self.worker_process:
+                self.worker_process.join()
             self.worker_process = None
-            self.RegButton_Papyrus.setText("START PAPYRUS MONITORING")
-            self.RegButton_Papyrus.setStyleSheet("color: black; background: rgb(45, 237, 138); border-radius: 10px")
+            self.papyrus_button.setText("START PAPYRUS MONITORING")
+            self.papyrus_button.setStyleSheet("""
+                QPushButton {
+                    color: black;
+                    background: rgb(45, 237, 138);
+                    border-radius: 10px;
+                    border: 1px solid black;
+                    font-weight: bold;
+                }
+            """)
         self.is_worker_running = not self.is_worker_running
 
-    def update_text_window(self):
+    def update_output_text_box_papyrus_watcher(self):
         while not self.result_queue.empty():
-            queue_result = self.result_queue.get()  # papyrus_result
-            old_papyrus_text = self.TXT_Window.toPlainText()
+            queue_result = self.result_queue.get()
             new_papyrus_text, new_dump_count = queue_result[:2]
+            old_papyrus_text = self.output_text_box.toPlainText()
             old_dump_count = next((line for line in old_papyrus_text.split("\n") if "DUMPS" in line), None)
             if old_dump_count and new_dump_count > int(old_dump_count.split(" : ")[1]):
                 play_sound("classic_error.wav")
                 time.sleep(3)
-            self.TXT_Window.setPlainText(new_papyrus_text)
-
-    def select_folder_scan(self):
-        SCAN_folder = QFileDialog.getExistingDirectory()
-        if SCAN_folder:
-            self.Box_SelectedScan.setText(SCAN_folder)
-            CMain.yaml_settings("CLASSIC Settings.yaml", f"CLASSIC_Settings.SCAN Custom Path", SCAN_folder)
-            self.Box_SelectedScan.setStyleSheet("color: black; font-family: Yu Gothic")
-
-    def select_folder_mods(self):
-        MODS_folder = QFileDialog.getExistingDirectory()
-        if MODS_folder:
-            self.Box_SelectedMods.setText(MODS_folder)
-            CMain.yaml_settings("CLASSIC Settings.yaml", f"CLASSIC_Settings.MODS Folder Path", MODS_folder)
-            self.Box_SelectedMods.setStyleSheet("color: black; font-family: Yu Gothic")
-
-    def select_folder_ini(self):
-        INI_folder = QFileDialog.getExistingDirectory()  # QFileDialog.getOpenFileName(filter="*.ini")
-        if INI_folder:
-            QtWidgets.QMessageBox.information(self, "New INI Path Set", "You have set the new path to: \n" + INI_folder)
-            CMain.yaml_settings("CLASSIC Settings.yaml", f"CLASSIC_Settings.INI Folder Path", INI_folder)
-
-        # ================== POP-UPS / WARNINGS =====================
-
-    @staticmethod
-    def open_settings():
-        settings_file = "CLASSIC Settings.yaml"
-        if platform.system() == "Windows":
-            subprocess.run(["start", "", settings_file], shell=True)
-        else:
-            subprocess.Popen(["xdg-open", settings_file])
-
-    def help_popup_main(self):
-        help_popup_text = CMain.yaml_settings("CLASSIC Data/databases/CLASSIC Main.yaml", "CLASSIC_Interface.help_popup_main")
-        popup = custom_popup_window(self, title="NEED HELP?", text=help_popup_text, height=450, callback="https://discord.com/invite/7ZZbrsGQh4")
-        popup.exec()
-
-    def update_popup(self):
-        update_popup_text = CMain.yaml_settings("CLASSIC Data/databases/CLASSIC Main.yaml", "CLASSIC_Interface.update_popup_text")
-        if self.is_uptodate:
-            popup = custom_popup_window(self, title="CLASSIC UPDATE", text="You have the latest version of CLASSIC!")
-            popup.exec()
-        else:
-            popup = custom_popup_window(self, title="CLASSIC UPDATE", text=update_popup_text, callback="https://github.com/evildarkarchon/CLASSIC-Fallout4/releases/latest")
-            popup.exec()
-
-
-class UiCLASSICBackups(QtWidgets.QMainWindow):
-    def __init__(self):
-        super().__init__()
-
-        # BACKGROUND CONFIG | UiCLASSICBackups
-        image_path = "CLASSIC Data/graphics/background.png"
-        self.background_label = QtWidgets.QLabel(self)
-        self.setCentralWidget(self.background_label)
-        pixmap = QtGui.QPixmap(image_path)
-        self.background_label.setPixmap(pixmap)
-        self.background_label.setScaledContents(True)
-
-        # TABS / SCREENS | UiCLASSICBackups
-        self.RegButton_TabMain = custom_push_button(self, QtCore.QRect(0, 0, 215, 48), "RegButton_TabMain", "MAIN OPTIONS", bold_11, "", self.open_tab_mainwin)
-        self.RegButton_TabMain.setStyleSheet("color: white; background: rgba(10, 10, 10, 0.90); border-radius: 0px; border : 2px dashed white; font-size: 15px")
-
-        self.RegButton_TabBackups = custom_push_button(self, QtCore.QRect(215, 0, 215, 48), "RegButton_TabBackups", "FILE BACKUP", bold_11, "")
-        self.RegButton_TabBackups.setStyleSheet("color: white; background: rgba(25, 25, 25, 0.90); border-radius: 0px; border : 2px solid white; font-size: 15px")
-
-        self.DropDown_GameSelect = custom_hover_button(self, QtCore.QRect(430, 0, 220, 48), "DropDown_GameSelect", "FALLOUT 4", bold_11, "")
-        self.DropDown_GameSelect.setStyleSheet("color: white; background: rgba(240, 63, 40, 0.75); border-radius: 0px; border : 2px solid white; font-size: 15px")
-
-        # EXPLANATIONS | UiCLASSICBackups
-        self.LBL_Backup = custom_label(self, QtCore.QRect(30, 65, 590, 20), "BACKUP > Backup files from the game folder into the CLASSIC Backup folder.", normal_11, "LBL_Backup")
-        self.LBL_Restore = custom_label(self, QtCore.QRect(30, 85, 590, 20), "RESTORE > Restore file backup from the CLASSIC Backup folder into the game folder.", normal_11, "LBL_Restore")
-        self.LBL_Remove = custom_label(self, QtCore.QRect(30, 105, 590, 20), "REMOVE > Remove files only from the game folder without removing existing backups.", normal_11, "LBL_Remove")
-
-        # SCRIPT EXTENDER BACKUP / RESTORE / REMOVE | UiCLASSICBackups
-        self.Line_Sep_XSE = custom_frame(self, QtCore.QRect(30, 130, 590, 20), QtWidgets.QFrame.Shape.HLine, QtWidgets.QFrame.Shadow.Sunken, "Line_Sep_XSE")
-        self.LBL_XSE = custom_label(self, QtCore.QRect(265, 160, 590, 20), "SCRIPT EXTENDER", bold_11, "LBL_XSE")
-
-        self.RegButton_BackupXSE = custom_push_button(self, QtCore.QRect(30, 205, 180, 48), "RegButton_BackupXSE", "BACKUP XSE", bold_11, "", partial(self.classic_files_manage, "Backup XSE", "BACKUP"))
-        self.RegButton_RestoreXSE = custom_push_button(self, QtCore.QRect(235, 205, 180, 48), "RegButton_RestoreXSE", "RESTORE XSE", bold_11, "", partial(self.classic_files_manage, "Backup XSE", "RESTORE"))
-        self.RegButton_RestoreXSE.setStyleSheet("color: white; border-radius: 10px; border : 2px solid rgb(250, 250, 250)")
-        self.RegButton_RemoveXSE = custom_push_button(self, QtCore.QRect(440, 205, 180, 48), "RegButton_RemoveXSE", "REMOVE XSE", bold_11, "", partial(self.classic_files_manage, "Backup XSE", "REMOVE"))
-        self.RegButton_RemoveXSE.setStyleSheet("color: white; border-radius: 10px; border : 2px solid  rgb(240, 63, 40)")
-
-        if os.path.isdir("CLASSIC Backup/Game Files/Backup XSE") and any(Path("CLASSIC Backup/Game Files/Backup XSE").iterdir()):
-            self.RegButton_BackupXSE.setStyleSheet("color: black; background: rgb(45, 237, 138); border-radius: 10px; border : 2px solid black")
-            self.RegButton_RestoreXSE.setEnabled(True)
-            self.RegButton_RestoreXSE.setStyleSheet("color: black; background: rgb(250, 250, 250); border-radius: 10px; border : 2px solid black;")
-        else:
-            self.RegButton_BackupXSE.setStyleSheet("color: white; border-radius: 10px; border : 2px solid rgb(45, 237, 138)")
-            self.RegButton_RestoreXSE.setEnabled(False)
-
-        # RESHADE BACKUP / RESTORE / REMOVE | UiCLASSICBackups
-        self.Line_Sep_RESHADE = custom_frame(self, QtCore.QRect(30, 280, 590, 20), QtWidgets.QFrame.Shape.HLine, QtWidgets.QFrame.Shadow.Sunken, "Line_Sep_RESHADE")
-        self.LBL_RESHADE = custom_label(self, QtCore.QRect(295, 310, 590, 20), "RESHADE", bold_11, "LBL_RESHADE")
-
-        self.RegButton_BackupRESHADE = custom_push_button(self, QtCore.QRect(30, 355, 180, 48), "RegButton_BackupRESHADE", "BACKUP RESHADE", bold_11, "", partial(self.classic_files_manage, "Backup RESHADE", "BACKUP"))
-        self.RegButton_RestoreRESHADE = custom_push_button(self, QtCore.QRect(235, 355, 180, 48), "RegButton_RestoreRESHADE", "RESTORE RESHADE", bold_11, "", partial(self.classic_files_manage, "Backup RESHADE", "RESTORE"))
-        self.RegButton_RestoreRESHADE.setStyleSheet("color: white; border-radius: 10px; border : 2px solid rgb(250, 250, 250)")
-        self.RegButton_RemoveRESHADE = custom_push_button(self, QtCore.QRect(440, 355, 180, 48), "RegButton_RemoveRESHADE", "REMOVE RESHADE", bold_11, "", partial(self.classic_files_manage, "Backup RESHADE", "REMOVE"))
-        self.RegButton_RemoveRESHADE.setStyleSheet("color: white; border-radius: 10px; border : 2px solid  rgb(240, 63, 40)")
-
-        if os.path.isdir("CLASSIC Backup/Game Files/Backup RESHADE") and any(Path("CLASSIC Backup/Game Files/Backup RESHADE").iterdir()):
-            self.RegButton_BackupRESHADE.setStyleSheet("color: black; background: rgb(45, 237, 138); border-radius: 10px; border : 2px solid black")
-            self.RegButton_RestoreRESHADE.setEnabled(True)
-            self.RegButton_RestoreRESHADE.setStyleSheet("color: black; background: rgb(250, 250, 250); border-radius: 10px; border : 2px solid black;")
-        else:
-            self.RegButton_BackupRESHADE.setStyleSheet("color: white; border-radius: 10px; border : 2px solid rgb(45, 237, 138)")
-            self.RegButton_RestoreRESHADE.setEnabled(False)
-
-        # VULKAN BACKUP / RESTORE / REMOVE | UiCLASSICBackups
-        self.Line_Sep_VULKAN = custom_frame(self, QtCore.QRect(30, 430, 590, 20), QtWidgets.QFrame.Shape.HLine, QtWidgets.QFrame.Shadow.Sunken, "Line_Sep_VULKAN")
-        self.LBL_VULKAN = custom_label(self, QtCore.QRect(260, 460, 590, 20), "VULKAN RENDERER", bold_11, "LBL_VULKAN")
-
-        self.RegButton_BackupVULKAN = custom_push_button(self, QtCore.QRect(30, 505, 180, 48), "RegButton_BackupVULKAN", "BACKUP VULKAN", bold_11, "", partial(self.classic_files_manage, "Backup VULKAN", "BACKUP"))
-        self.RegButton_RestoreVULKAN = custom_push_button(self, QtCore.QRect(235, 505, 180, 48), "RegButton_RestoreVULKAN", "RESTORE VULKAN", bold_11, "", partial(self.classic_files_manage, "Backup VULKAN", "RESTORE"))
-        self.RegButton_RestoreVULKAN.setStyleSheet("color: white; border-radius: 10px; border : 2px solid rgb(250, 250, 250)")
-        self.RegButton_RemoveVULKAN = custom_push_button(self, QtCore.QRect(440, 505, 180, 48), "RegButton_RemoveVULKAN", "REMOVE VULKAN", bold_11, "", partial(self.classic_files_manage, "Backup VULKAN", "REMOVE"))
-        self.RegButton_RemoveVULKAN.setStyleSheet("color: white; border-radius: 10px; border : 2px solid  rgb(240, 63, 40)")
-
-        if os.path.isdir("CLASSIC Backup/Game Files/Backup VULKAN") and any(Path("CLASSIC Backup/Game Files/Backup VULKAN").iterdir()):
-            self.RegButton_BackupVULKAN.setStyleSheet("color: black; background: rgb(45, 237, 138); border-radius: 10px; border : 2px solid black")
-            self.RegButton_RestoreVULKAN.setEnabled(True)
-            self.RegButton_RestoreVULKAN.setStyleSheet("color: black; background: rgb(250, 250, 250); border-radius: 10px; border : 2px solid black;")
-        else:
-            self.RegButton_BackupVULKAN.setStyleSheet("color: white; border-radius: 10px; border : 2px solid rgb(45, 237, 138)")
-            self.RegButton_RestoreVULKAN.setEnabled(False)
-
-        # ENB BACKUP / RESTORE / REMOVE | UiCLASSICBackups
-        self.Line_Sep_ENB = custom_frame(self, QtCore.QRect(30, 580, 590, 20), QtWidgets.QFrame.Shape.HLine, QtWidgets.QFrame.Shadow.Sunken, "Line_Sep_ENB")
-        self.LBL_ENB = custom_label(self, QtCore.QRect(200, 610, 590, 20), "ENHANCED NATURAL BEAUTY ( ENB )", bold_11, "LBL_ENB")
-
-        self.RegButton_BackupENB = custom_push_button(self, QtCore.QRect(30, 655, 180, 48), "RegButton_BackupENB", "BACKUP ENB", bold_11, "", partial(self.classic_files_manage, "Backup ENB", "BACKUP"))
-        self.RegButton_RestoreENB = custom_push_button(self, QtCore.QRect(235, 655, 180, 48), "RegButton_RestoreENB", "RESTORE ENB", bold_11, "", partial(self.classic_files_manage, "Backup ENB", "RESTORE"))
-        self.RegButton_RestoreENB.setStyleSheet("color: white; border-radius: 10px; border : 2px solid rgb(250, 250, 250)")
-        self.RegButton_RemoveENB = custom_push_button(self, QtCore.QRect(440, 655, 180, 48), "RegButton_RemoveENB", "REMOVE ENB", bold_11, "", partial(self.classic_files_manage, "Backup ENB", "REMOVE"))
-        self.RegButton_RemoveENB.setStyleSheet("color: white; border-radius: 10px; border : 2px solid  rgb(240, 63, 40)")
-
-        if os.path.isdir("CLASSIC Backup/Game Files/Backup ENB") and any(Path("CLASSIC Backup/Game Files/Backup ENB").iterdir()):
-            self.RegButton_BackupENB.setStyleSheet("color: black; background: rgb(45, 237, 138); border-radius: 10px; border : 2px solid black")
-            self.RegButton_RestoreENB.setEnabled(True)
-            self.RegButton_RestoreENB.setStyleSheet("color: black; background: rgb(250, 250, 250); border-radius: 10px; border : 2px solid black;")
-        else:
-            self.RegButton_BackupENB.setStyleSheet("color: white; border-radius: 10px; border : 2px solid rgb(45, 237, 138)")
-            self.RegButton_RestoreENB.setEnabled(False)
-
-        # BOTTOM | UiCLASSICBackups
-
-        # Button - HELP | UiCLASSICBackups
-        self.RegButton_Help = custom_push_button(self, QtCore.QRect(20, 770, 110, 30), "RegButton_Help", "HELP", normal_11, "", self.help_popup_backup)
-        # Button - BACKUPS FOLDER | UiCLASSICBackups
-        self.RegButton_OpenBackup = custom_push_button(self, QtCore.QRect(215, 770, 220, 30), "RegButton_OpenBackup", "OPEN CLASSIC BACKUPS", bold_11, "", self.open_backup_folder)
-        # Button - EXIT | UiCLASSICBackups
-        self.RegButton_Exit = custom_push_button(self, QtCore.QRect(520, 770, 110, 30), "RegButton_Exit", "EXIT", normal_11, "", lambda: QtWidgets.QApplication.quit())
-        # Text Box - SHARED | UiCLASSICBackups
-        self.TXT_Window = custom_text_box(self, QtCore.QRect(20, 810, 610, 120), "Crash Log Auto Scanner & Setup Integrity Checker | Made by: Poet \nContributors: evildarkarchon | kittivelae | AtomicFallout757")
-
-        QtCore.QMetaObject.connectSlotsByName(self)
-
-    @staticmethod
-    def open_tab_mainwin():
-        screen_switch.setCurrentIndex(screen_switch.currentIndex() - 1)
-
-    @staticmethod
-    def open_backup_folder():
-        if platform.system() == "Windows":
-            os.system('explorer "CLASSIC Backup\\Game Files"')
-        else:
-            os.system('xdg-open "CLASSIC Backup\\Game Files"')
-
-    def help_popup_backup(self):
-        help_popup_text = CMain.yaml_settings("CLASSIC Data/databases/CLASSIC Main.yaml", "CLASSIC_Interface.help_popup_backup")
-        popup = custom_popup_window(self, title="NEED HELP?", text=help_popup_text, height=450, callback="https://discord.com/invite/7ZZbrsGQh4")
-        popup.exec()
-
-    # noinspection PyMethodMayBeStatic
-    def classic_files_manage(self, selected_list, selected_mode="BACKUP"):
-        list_name = selected_list.split(" ", 1)
-
-        try:
-            CGame.game_files_manage(selected_list, selected_mode)
-            if selected_mode == "BACKUP":
-                exec(f"self.RegButton_Restore{list_name[1]}.setEnabled(True)")
-        except PermissionError:
-            print(f"❌ ERROR : UNABLE TO ACCESS FILES FROM YOUR GAME FOLDER.\n   TO RESOLVE THIS PROBLEM, PLEASE RUN CLASSIC IN ADMIN MODE \n")
-
+            self.output_text_box.setPlainText(new_papyrus_text)
 
 if __name__ == "__main__":
-    CMain.main_generate_required()
-    print(CMain.yaml_settings("CLASSIC Data/databases/CLASSIC Main.yaml", "CLASSIC_Interface.start_message"))
-    classic_ver = CMain.yaml_settings("CLASSIC Data/databases/CLASSIC Main.yaml", "CLASSIC_Info.version")
+    app = QApplication(sys.argv)
 
-    app = QtWidgets.QApplication(sys.argv)
-    if platform.system() == "Windows":
-        app.setStyle("windowsvista")
-
-    # Add widgets of other "tabs" through function calls, not here.
-    screen_switch = QtWidgets.QStackedWidget()
-    screen_main = UiCLASSICMainWin()
-    screen_switch.addWidget(screen_main)
-
-    # Stuff here acts as defaults, no need to set anywhere else.
-    screen_switch.setWindowTitle(f"Crash Log Auto Scanner & Setup Integrity Checker | {classic_ver}")
-    screen_switch.setWindowIcon(QIcon("CLASSIC Data/graphics/CLASSIC.ico"))
-    screen_switch.setStyleSheet("font-family: Yu Gothic; font-size: 13px")
-    screen_switch.setObjectName("CLASSIC_MainWin")
-    screen_switch.setFixedWidth(650)
-    screen_switch.setFixedHeight(950)
-
-    # Main window to the middle of the screen with 75px top bias.
-    screen_geometry = QtWidgets.QApplication.primaryScreen().geometry()
-    x = (screen_geometry.width() - screen_switch.width()) // 2
-    y = (screen_geometry.height() - 75 - screen_switch.height()) // 2
-    screen_switch.move(x, y)
-    screen_switch.show()
-    sys.exit(app.exec())
+    try:
+        window = MainWindow()
+        window.show()
+        sys.exit(app.exec())
+    except Exception as e:
+        error_text = traceback.format_exc()
+        show_exception_box(error_text)
